Time: 26 ms (13.88%) | Memory: 47.9 MB (58.23%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.java
@@ -2,7 +2,7 @@ class Solution {
     static boolean isValidAns(int[] piles,int h,int mid){
         long hours = 0;
         for(int i= 0; i< piles.length; i++){
-         hours += (long) Math.ceil((double) piles[i] /  mid);
+         hours +=  Math.ceil((double) piles[i] /  mid);
          if(hours > h){
             return false;
          }
